Move classes to TestCommon namespace
