+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.IntegrationTests/Fixtures/TestCommon/DatabricksWarehouseManager.cs
@@ -16,7 +16,7 @@ using System.Net.Http.Headers;
 using System.Net.Http.Json;
 using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements;
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Fixtures;
+namespace Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Fixtures.TestCommon;
 
 /// <summary>
 /// A manager for managing Databricks SQL schemas and tables from integration tests.
