+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.IntegrationTests/Fixtures/TestCommon/DatabricksWarehouseSettings.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Fixtures
+namespace Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Fixtures.TestCommon
 {
     /// <summary>
     /// Settings used by <see cref="DatabricksWarehouseManager"/>.
